Add integration test for feedback
