+++ app/v2/definitions/filename_mapper.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 class FileNameMapperBase(ABC):
 
     @abstractmethod
-    def get_output_type(self, filename: str) -> str:
+    def get_output_type(self, filename: str, submission_type: str) -> str:
         pass
